requirements.txt
datasets[audio]==3.6.0
transformers==4.26.1
accelerate
evaluate
jiwer
tensorboard
gradio
torchaudio
matplotlib